requirements.txt
streamlit
langchain
langchain-groq
langchain-huggingface
langchain-chroma
langchain-community
python-dotenv
pypdf
sentence-transformers
chromadb